requirements.txt
flask
faster-whisper
TTS
google-generativeai
python-dotenv
imageio-ffmpeg
